skills/transcribe/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcribe-minimal"
version = "0.1.0"
description = "Minimal FunASR-first transcription pipeline"
requires-python = ">=3.11"
dependencies = [
  "dashscope",
  "PyYAML>=6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
